armv8: Remove unnecessary sign extensions
After loading a single byte with ldrb and immediately moving
it into a vector register with dup .8b, there's no point in
sign extending it to 64 bit inbetween.

Similarly in the second case, when adding a number of pixels,
there's no need to sign extend between additions, just do the
additions using the 32 bit register names.

Change-Id: Ie37f47b93138643ffd05169a3792bc39967dbea3
Signed-off-by: mydongistiny <jaysonedson@gmail.com>